refactor(runner): extract config generation into standalone `runner config` command
- `runner build` now only builds rootfs + snapshot and outputs hashes to stdout
- New `runner config` command takes --rootfs-hash and --snapshot-hash to generate runner.yaml
- `run_snapshot` returns (hash, SnapshotConfig) tuple so callers can access the hash
- CI workflows updated: deploy-runner-start uses `runner config` instead of re-running build

Closes #3159
