+++ crates/runner/src/cmd/mod.rs
@@ -1,5 +1,6 @@
 mod benchmark;
 mod build;
+mod config;
 mod doctor;
 mod gc;
 mod kill;
@@ -12,6 +13,7 @@ mod submit;
 
 pub use benchmark::{BenchmarkArgs, run_benchmark};
 pub use build::{BuildArgs, run_build};
+pub use config::{ConfigArgs, run_config};
 pub use doctor::{DoctorArgs, run_doctor};
 pub use gc::{GcArgs, run_gc};
 pub use kill::{KillArgs, run_kill};
